#913 added test to utils_test.py
